backend/pyproject.toml
[project]
name = "market-wizard"
version = "0.1.0"
description = "Market and product analyzer using SSR (Semantic Similarity Rating) methodology"
# readme = "README.md"  # README is in root directory
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Paweł"}
]
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy>=2.0",
    "aiosqlite",
    "polars>=1.0",
    "numpy>=2.0",
    "pydantic>=2.0",
    "pydantic-settings",
    "google-genai>=1.50",  # New Gemini SDK
    "openai>=1.50",
    "sentence-transformers>=3.0",  # For local BGE-M3 embeddings
    "gradio>=5.0",
    "httpx>=0.27",
    "matplotlib>=3.8.0",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio",
    "ruff",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["app"]
